+++ league/management/commands/scheduleWorse.py
@@ -106,15 +106,36 @@ class Command(BaseCommand):
         return confTeams
     def getGames(self, team):
         games = Game.objects.filter(Q(homeTeam=team) | Q(awayTeam=team))
-        homeTeams = games.values_list('homeTeam', flat=True).distinct()
-        awayTeams = games.values_list('awayTeam', flat=True).distinct()
-        
+        homeTeams = games.values_list('homeTeam', flat=True)
+        awayTeams = games.values_list('awayTeam', flat=True)
         
     def checkConfGames(self, team):
         confTeams = self.getConfTeams(team)
         
         
-
+        games = Game.objects.filter(Q(homeTeam=team) | Q(awayTeam=team))
+        
+        confTeams = Team.objects.filter(Q(conference=team.conference) & ~Q(division=team.division) & ~Q(t_id=team.t_id))
+        
+        
+        confGamesDiv = games.filter(Q(homeTeam__conference = team.conference) | Q(awayTeam__conference = team.conference))
+        confGames = confGamesDiv.filter(~Q(homeTeam__division = team.division) | ~Q(awayTeam__division = team.division))
+        
+        confTeamsHome = confGames.values_list('homeTeam', flat=True).distinct()
+        confTeamsAway = confGames.values_list('awayTeam', flat=True).distinct()
+        confTeamsPlayed = list(set(list(confTeamsHome) + list(confTeamsAway)))
+        
+        try:
+            confTeamsPlayed.remove(team)
+        except ValueError:
+            pass
+        
+        confTeams = list(confTeams)
+        for team in confTeamsPlayed:
+            if team in confTeams:
+                confTeams.remove(team)
+                
+        return confTeams
         
         
     
@@ -246,9 +267,3 @@ class Command(BaseCommand):
         #     for divTeam in divOppenents:
         #         self.scheduleDivOppenent(teams[0], divTeam, numberedGamesList)
             
-        
-
-        
-        
-        
-
